requirements.txt
azure-functions==1.7.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
urllib3==1.26.4
